Re-write of Indexer API wrapper Re-write of New Show seearch function Re-write of Existing show search helper function for determining indexer/indexerid Massive code cleanup and more bugs found and fixed Indexer code fully modualized for future proofing

+++ sickbeard/processTV.py
@@ -36,11 +36,14 @@ from sickbeard import failedProcessor
 from lib.unrar2 import RarFile, RarInfo
 from lib.unrar2.rar_exceptions import *
 
-def logHelper (logMessage, logLevel=logger.MESSAGE):
+
+def logHelper(logMessage, logLevel=logger.MESSAGE):
     logger.log(logMessage, logLevel)
     return logMessage + u"\n"
 
-def processDir(dirName, nzbName=None, process_method=None, force=False, is_priority=None, failed=False, type="auto", indexer="auto"):
+
+def processDir(dirName, nzbName=None, process_method=None, force=False, is_priority=None, failed=False, type="auto",
+               indexer="auto"):
     """
     Scans through the files in dirName and processes whatever media files it finds
 
@@ -72,7 +75,9 @@ def processDir(dirName, nzbName=None, process_method=None, force=False, is_prior
 
     # if we didn't find a real dir then quit
     if not ek.ek(os.path.isdir, dirName):
-        returnStr += logHelper(u"Unable to figure out what folder to process. If your downloader and Sick Beard aren't on the same PC make sure you fill out your TV download dir in the config.", logger.DEBUG)
+        returnStr += logHelper(
+            u"Unable to figure out what folder to process. If your downloader and Sick Beard aren't on the same PC make sure you fill out your TV download dir in the config.",
+            logger.DEBUG)
         return returnStr
 
     path, dirs, files = get_path_dir_files(dirName, nzbName, type)
@@ -128,26 +133,28 @@ def processDir(dirName, nzbName=None, process_method=None, force=False, is_prior
             #Don't Link media when the media is extracted from a rar in the same path
             if process_method in ('hardlink', 'symlink') and videoInRar:
                 process_media(processPath, videoInRar, nzbName, 'move', force, is_priority, indexer)
-                process_media(processPath, set(videoFiles) - set(videoInRar), nzbName, process_method, force, is_priority, indexer)
+                process_media(processPath, set(videoFiles) - set(videoInRar), nzbName, process_method, force,
+                              is_priority, indexer)
                 delete_files(processPath, rarContent)
             else:
                 process_media(processPath, videoFiles, nzbName, process_method, force, is_priority, indexer)
 
                 #Delete all file not needed
                 if process_method != "move" or not process_result \
-                or type=="manual": #Avoid to delete files if is Manual PostProcessing
+                        or type == "manual":  #Avoid to delete files if is Manual PostProcessing
                     continue
 
                 delete_files(processPath, notwantedFiles)
 
                 if process_method == "move" and \
-                ek.ek(os.path.normpath, processPath) != ek.ek(os.path.normpath, sickbeard.TV_DOWNLOAD_DIR):
+                                ek.ek(os.path.normpath, processPath) != ek.ek(os.path.normpath,
+                                                                              sickbeard.TV_DOWNLOAD_DIR):
                     delete_dir(processPath)
 
     return returnStr
 
-def validateDir(path, dirName, nzbNameOriginal, failed):
 
+def validateDir(path, dirName, nzbNameOriginal, failed):
     global process_result, returnStr
 
     returnStr += logHelper(u"Processing folder " + dirName, logger.DEBUG)
@@ -156,10 +163,12 @@ def validateDir(path, dirName, nzbNameOriginal, failed):
         returnStr += logHelper(u"The directory name indicates it failed to extract.", logger.DEBUG)
         failed = True
     elif ek.ek(os.path.basename, dirName).startswith('_UNDERSIZED_'):
-        returnStr += logHelper(u"The directory name indicates that it was previously rejected for being undersized.", logger.DEBUG)
+        returnStr += logHelper(u"The directory name indicates that it was previously rejected for being undersized.",
+                               logger.DEBUG)
         failed = True
     elif ek.ek(os.path.basename, dirName).startswith('_UNPACK_'):
-        returnStr += logHelper(u"The directory name indicates that this release is in the process of being unpacked.", logger.DEBUG)
+        returnStr += logHelper(u"The directory name indicates that this release is in the process of being unpacked.",
+                               logger.DEBUG)
 
     if failed:
         process_failed(os.path.join(path, dirName), nzbNameOriginal)
@@ -169,8 +178,12 @@ def validateDir(path, dirName, nzbNameOriginal, failed):
     myDB = db.DBConnection()
     sqlResults = myDB.select("SELECT * FROM tv_shows")
     for sqlShow in sqlResults:
-        if dirName.lower().startswith(ek.ek(os.path.realpath, sqlShow["location"]).lower()+os.sep) or dirName.lower() == ek.ek(os.path.realpath, sqlShow["location"]).lower():
-            returnStr += logHelper(u"You're trying to post process an episode that's already been moved to its show dir, skipping", logger.ERROR)
+        if dirName.lower().startswith(
+                        ek.ek(os.path.realpath, sqlShow["location"]).lower() + os.sep) or dirName.lower() == ek.ek(
+                os.path.realpath, sqlShow["location"]).lower():
+            returnStr += logHelper(
+                u"You're trying to post process an episode that's already been moved to its show dir, skipping",
+                logger.ERROR)
             return False
 
     # Get the videofile list for the next checks
@@ -211,8 +224,8 @@ def validateDir(path, dirName, nzbNameOriginal, failed):
 
     return False
 
-def unRAR(path, rarFiles, force):
 
+def unRAR(path, rarFiles, force):
     global process_result, returnStr
 
     unpacked_files = []
@@ -232,27 +245,29 @@ def unRAR(path, rarFiles, force):
                 skip_file = False
                 for file_in_archive in [os.path.basename(x.filename) for x in rar_handle.infolist() if not x.isdir]:
                     if already_postprocessed(path, file_in_archive, force):
-                        returnStr += logHelper(u"Archive file already post-processed, extraction skipped: " + file_in_archive, logger.DEBUG)
+                        returnStr += logHelper(
+                            u"Archive file already post-processed, extraction skipped: " + file_in_archive,
+                            logger.DEBUG)
                         skip_file = True
                         break
 
                 if skip_file:
                     continue
 
-                rar_handle.extract(path = path, withSubpath = False, overwrite = False)
+                rar_handle.extract(path=path, withSubpath=False, overwrite=False)
                 unpacked_files += [os.path.basename(x.filename) for x in rar_handle.infolist() if not x.isdir]
                 del rar_handle
             except Exception, e:
-                 returnStr += logHelper(u"Failed Unrar archive " + archive + ': ' + ex(e), logger.ERROR)
-                 process_result = False
-                 continue
+                returnStr += logHelper(u"Failed Unrar archive " + archive + ': ' + ex(e), logger.ERROR)
+                process_result = False
+                continue
 
         returnStr += logHelper(u"UnRar content: " + str(unpacked_files), logger.DEBUG)
 
     return unpacked_files
 
-def already_postprocessed(dirName, videofile, force):
 
+def already_postprocessed(dirName, videofile, force):
     global returnStr
 
     if force:
@@ -266,7 +281,8 @@ def already_postprocessed(dirName, videofile, force):
     myDB = db.DBConnection()
     sqlResult = myDB.select("SELECT * FROM tv_episodes WHERE release_name = ?", [dirName])
     if sqlResult:
-        returnStr += logHelper(u"You're trying to post process a dir that's already been processed, skipping", logger.DEBUG)
+        returnStr += logHelper(u"You're trying to post process a dir that's already been processed, skipping",
+                               logger.DEBUG)
         return True
 
     # This is needed for video whose name differ from dirName
@@ -275,7 +291,8 @@ def already_postprocessed(dirName, videofile, force):
 
     sqlResult = myDB.select("SELECT * FROM tv_episodes WHERE release_name = ?", [videofile.rpartition('.')[0]])
     if sqlResult:
-        returnStr += logHelper(u"You're trying to post process a video that's already been processed, skipping", logger.DEBUG)
+        returnStr += logHelper(u"You're trying to post process a video that's already been processed, skipping",
+                               logger.DEBUG)
         return True
 
     #Needed if we have downloaded the same episode @ different quality
@@ -285,13 +302,14 @@ def already_postprocessed(dirName, videofile, force):
     search_sql += " and history.resource LIKE ?"
     sqlResult = myDB.select(search_sql, [u'%' + videofile])
     if sqlResult:
-        returnStr += logHelper(u"You're trying to post process a video that's already been processed, skipping", logger.DEBUG)
+        returnStr += logHelper(u"You're trying to post process a video that's already been processed, skipping",
+                               logger.DEBUG)
         return True
 
     return False
 
-def process_media(processPath, videoFiles, nzbName, process_method, force, is_priority, indexer):
 
+def process_media(processPath, videoFiles, nzbName, process_method, force, is_priority, indexer):
     global process_result, returnStr
 
     for cur_video_file in videoFiles:
@@ -314,14 +332,15 @@ def process_media(processPath, videoFiles, nzbName, process_method, force, is_pr
         if process_result:
             returnStr += logHelper(u"Processing succeeded for " + cur_video_file_path)
         else:
-            returnStr += logHelper(u"Processing failed for " + cur_video_file_path + ": " + process_fail_message, logger.WARNING)
+            returnStr += logHelper(u"Processing failed for " + cur_video_file_path + ": " + process_fail_message,
+                                   logger.WARNING)
 
         #If something fail abort the processing on dir
         if not process_result:
             break
 
-def delete_files(processPath, notwantedFiles):
 
+def delete_files(processPath, notwantedFiles):
     global returnStr, process_result
 
     if not process_result:
@@ -333,30 +352,32 @@ def delete_files(processPath, notwantedFiles):
         cur_file_path = ek.ek(os.path.join, processPath, cur_file)
 
         if not ek.ek(os.path.isfile, cur_file_path):
-            continue #Prevent error when a notwantedfiles is an associated files
+            continue  #Prevent error when a notwantedfiles is an associated files
 
         returnStr += logHelper(u"Deleting file " + cur_file, logger.DEBUG)
 
-                #check first the read-only attribute
+        #check first the read-only attribute
         file_attribute = ek.ek(os.stat, cur_file_path)[0]
         if (not file_attribute & stat.S_IWRITE):
             # File is read-only, so make it writeable
             returnStr += logHelper(u"Changing ReadOnly Flag for file " + cur_file, logger.DEBUG)
             try:
-                ek.ek(os.chmod,cur_file_path,stat.S_IWRITE)
+                ek.ek(os.chmod, cur_file_path, stat.S_IWRITE)
             except OSError, e:
-                returnStr += logHelper(u"Cannot change permissions of " + cur_file_path + ': ' + e.strerror, logger.DEBUG)
+                returnStr += logHelper(u"Cannot change permissions of " + cur_file_path + ': ' + e.strerror,
+                                       logger.DEBUG)
         try:
             ek.ek(os.remove, cur_file_path)
         except OSError, e:
             returnStr += logHelper(u"Unable to delete file " + cur_file + ': ' + e.strerror, logger.DEBUG)
 
-def delete_dir(processPath):
 
+def delete_dir(processPath):
     global returnStr
 
     if not ek.ek(os.listdir, processPath) == []:
-        returnStr += logHelper(u"Skipping Deleting folder " + processPath + ' because some files was not deleted/processed', logger.DEBUG)
+        returnStr += logHelper(
+            u"Skipping Deleting folder " + processPath + ' because some files was not deleted/processed', logger.DEBUG)
         return
 
     returnStr += logHelper(u"Deleting folder " + processPath, logger.DEBUG)
@@ -366,15 +387,16 @@ def delete_dir(processPath):
     except (OSError, IOError), e:
         returnStr += logHelper(u"Warning: unable to remove the folder " + processPath + ": " + ex(e), logger.WARNING)
 
-def get_path_dir_files(dirName, nzbName, type):
 
-    if dirName == sickbeard.TV_DOWNLOAD_DIR and not nzbName or type =="manual": #Scheduled Post Processing Active
+def get_path_dir_files(dirName, nzbName, type):
+    if dirName == sickbeard.TV_DOWNLOAD_DIR and not nzbName or type == "manual":  #Scheduled Post Processing Active
         #Get at first all the subdir in the dirName
         for path, dirs, files in ek.ek(os.walk, dirName):
             break
     else:
-        path, dirs = ek.ek(os.path.split, dirName) #Script Post Processing
-        if not nzbName is None and not nzbName.endswith('.nzb') and os.path.isfile(os.path.join(dirName, nzbName)): #For single torrent file without Dir
+        path, dirs = ek.ek(os.path.split, dirName)  #Script Post Processing
+        if not nzbName is None and not nzbName.endswith('.nzb') and os.path.isfile(
+                os.path.join(dirName, nzbName)):  #For single torrent file without Dir
             dirs = []
             files = [os.path.join(dirName, nzbName)]
         else:
@@ -382,6 +404,7 @@ def get_path_dir_files(dirName, nzbName, type):
             files = []
 
     return path, dirs, files
+
 
 def process_failed(dirName, nzbName):
     """Process a download that did not complete correctly"""
@@ -405,4 +428,6 @@ def process_failed(dirName, nzbName):
         if process_result:
             returnStr += logHelper(u"Failed Download Processing succeeded: (" + str(nzbName) + ", " + dirName + ")")
         else:
-            returnStr += logHelper(u"Failed Download Processing failed: (" + str(nzbName) + ", " + dirName + "): " + process_fail_message, logger.WARNING)
+            returnStr += logHelper(
+                u"Failed Download Processing failed: (" + str(nzbName) + ", " + dirName + "): " + process_fail_message,
+                logger.WARNING)
